requirements.txt
gradio>=4.0.0
transformers>=4.35.0
torch>=2.0.0
torchaudio>=2.0.0
librosa>=0.10.0
numpy>=1.24.0
matplotlib>=3.7.0
google-genai>=0.3.0
anthropic>=0.25.0
python-dotenv>=1.0.0
soundfile>=0.12.0